refactor: Fix version, update elvis config and variable renaming (#17)
* no-release: Fix id_token version but skip the release process

* Simplify maintenance of Elvis linter rules

* rebar3_lint: fix for consistent_variable_casing

* rebar3_lint: fix for no_trailing_whitespace

* rebar3_lint: "fix" for no_block_expressions

A proper fix would probably be to not avoid it,
but 1. that'd change code, which is not our
priority here, 2. that might not be easy to
achieve when macros are involved

* rebar3_lint: fix for used_ignored_variable

* rebar3_lint: fix for operator_spaces

* rebar3_lint: "fix" for no_debug_call

Instead of an actual fix we work around it by allowing
it; on the other hand I'm not sure why it exists in the
code but it's not our priority to check that now

* Make all Jw[<lowercase>] JW[<uppercase>] for readability

* Delete unwarranted push

---------

Co-authored-by: Paulo F. Oliveira <paulo.oliveira@kivra.com>